pyproject.toml
[tool.poetry]
name = "k3s-agent-operator"
version = "0.1.0"
description = ""
authors = ["Devin Young <devin@devinyoungweb.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "k3ao"}]

[tool.poetry.dependencies]
python = "^3.9"
kopf = "^1.36.0"
kubernetes = "^26.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
